Changing long/lat to BigDecimals. Fixing PostgreSQL dialect

+++ beacon-api/src/main/java/fr/woorib/beacon/data/BeaconEntry.java
@@ -1,5 +1,7 @@
 package fr.woorib.beacon.data;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Veryeld on 30/03/2017.
  */
@@ -7,9 +9,9 @@ public interface BeaconEntry {
 
     public Integer getUserId();
 
-    public Double getLongitude();
+    public BigDecimal getLongitude();
 
-    public Double getLatitude();
+    public BigDecimal getLatitude();
 
     public int getBeaconId();
