+++ beacon-web/src/main/java/fr/woorib/beacon/web/BeaconRest.java
@@ -1,5 +1,6 @@
 package fr.woorib.beacon.web;
 
+import java.math.BigDecimal;
 import java.util.List;
 import javax.inject.Inject;
 import org.springframework.stereotype.Controller;
@@ -34,8 +35,8 @@ public class BeaconRest {
   @RequestMapping("createBeacon")
   @ResponseBody
   public BeaconEntry createBeacon(@RequestParam(value="userId", required=true) Integer userId,
-                                  @RequestParam(value="longitude", required=true) Double longitude,
-                                  @RequestParam(value="latitude", required=true) Double latitude,
+                                  @RequestParam(value="longitude", required=true) BigDecimal longitude,
+                                  @RequestParam(value="latitude", required=true) BigDecimal latitude,
                                   Model model) {
     Integer beaconId = beaconService.setBeacon(userId, latitude, longitude);
     return beaconService.getBeacon(beaconId);
